+++ example/cli-template.txt
@@ -1,0 +1,10 @@
+-  Build program: 
+go build -o bridge-client
+
+
+- Start relayer: 
+./bridge-cli run --config ./cfg/config_evm-evm_1.json --latest
+
+
+- Deposit to bridge:
+./bridge-cli evm-cli erc20 deposit --url $SRC_GATEWAY --private-key $USER_1_PRIVATE_KEY --gas-price 25000000000 --amount 0.0001 --domain 1 --bridge $SRC_BRIDGE --recipient $USER_1_ADDR --resource $RESOURCE_ID --decimals 18
