+++ example/config.env
@@ -45,7 +45,7 @@ DST_TOKEN="0xb15e88ddDA2727c6938E592125F148EF2A30EBFA"
 
 USER_1_ADDR="0xF7208401386b1dc834561C92f33910E337bf00c0"
 USER_1_PRIVATE_KEY="bfafb7a7a2e90864303d477b71e1c1c64c2059a82f85c4706b6c6ceec8473959"
-USER_2_ADDR=""
-USER_2_PRIVATE_KEY=""
+USER_2_ADDR="0xb5d0E46cB174E0e1824dfA9f3837eabC8E5Df79d"
+USER_2_PRIVATE_KEY="9aa4cab962beb7fd544b03403a9c08ed4dd791ff5724e559e65b86e4ea24b7b5"
 USER_3_ADDR=""
 USER_3_PRIVATE_KEY=""
