+++ example/log_evaluate_old.txt
@@ -1,0 +1,73 @@
+2023/04/12 12:52:43 
+2023/04/12 12:52:43 ============================================================================
+2023/04/12 12:53:02 Step 0 (finish deposit): 18.611356744s
+2023/04/12 12:53:22 
+2023/04/12 12:53:22 ============================================================================
+2023/04/12 12:53:40 Step 0 (finish deposit): 17.625590988s
+2023/04/12 12:55:11 
+2023/04/12 12:55:11 ============================================================================
+2023/04/12 12:55:30 Step 0 (finish deposit): 19.242636768s
+2023/04/12 13:02:37 
+2023/04/12 13:02:37 ============================================================================
+2023/04/12 13:02:37 Step 0 (finish deposit): 22.984565807s
+2023/04/12 13:05:05 
+2023/04/12 13:05:05 ============================================================================
+2023/04/12 13:05:05 Step 0 (finish deposit): 18.609331763s
+2023/04/12 13:07:04 
+2023/04/12 13:07:04 ============================================================================
+2023/04/12 13:07:04 Step 0 (finish deposit): 18.85595365s
+2023/04/12 13:09:53 
+2023/04/12 13:09:53 ============================================================================
+2023/04/12 13:09:53 Step 0 (finish deposit): 18.42909979s
+2023/04/12 13:10:50 Debug T0a: 0001-01-01 00:00:00 +0000 UTC
+2023/04/12 13:17:54 Debug T0: 2023-04-12 13:17:54.155135218 +0000 UTC m=+0.016359196
+2023/04/12 13:18:01 Debug T0a: 2023-04-12 13:18:01.29619879 +0000 UTC m=+7.157422768
+2023/04/12 13:18:23 Debug T0: 2023-04-12 13:18:23.377370954 +0000 UTC m=+0.017165607
+2023/04/12 13:18:41 Debug T0a: 2023-04-12 13:18:41.105239847 +0000 UTC m=+17.745034500
+2023/04/12 13:18:42 Debug T1: 2023-04-12 13:18:42.29636946 +0000 UTC m=+55.199003316
+2023/04/12 13:18:44 Debug T2: 2023-04-12 13:18:44.625205041 +0000 UTC m=+57.527838997
+2023/04/12 13:19:26 Debug T0: 2023-04-12 13:19:26.622355746 +0000 UTC m=+0.016274095
+2023/04/12 13:19:50 Debug T0a: 2023-04-12 13:19:50.322884835 +0000 UTC m=+23.716803184
+2023/04/12 13:19:52 Debug T1: 2023-04-12 13:19:52.360741845 +0000 UTC m=+125.263375801
+2023/04/12 13:19:58 Debug T2: 2023-04-12 13:19:58.689202338 +0000 UTC m=+131.591836294
+2023/04/12 13:21:16 
+2023/04/12 13:21:16 ============================================================================
+2023/04/12 13:21:16 Debug T0: 2023-04-12 13:21:16.110670691 +0000 UTC m=+0.028934946
+2023/04/12 13:21:28 Debug T0a: 2023-04-12 13:21:28.486244304 +0000 UTC m=+12.404508559
+2023/04/12 13:22:29 Debug T1: 2023-04-12 13:22:29.234299461 +0000 UTC m=+82.276880649
+2023/04/12 13:22:43 Debug T2: 2023-04-12 13:22:43.585939493 +0000 UTC m=+96.628520681
+2023/04/12 13:24:54 
+2023/04/12 13:24:54 ============================================================================
+2023/04/12 13:24:54 Debug T0: 2023-04-12 13:24:54.343166993 +0000 UTC m=+0.039077465
+2023/04/12 13:25:12 Debug T0a: 2023-04-12 13:25:12.9025294 +0000 UTC m=+18.598439872
+2023/04/12 13:25:50 Debug T1: 2023-04-12 13:25:50.086407079 +0000 UTC m=+283.128988167
+2023/04/12 13:25:53 Debug T2: 2023-04-12 13:25:53.42227737 +0000 UTC m=+286.464858458
+2023/04/12 13:28:16 
+2023/04/12 13:28:16 ============================================================================
+2023/04/12 13:28:16 Debug T0: 2023-04-12 13:28:16.867928226 +0000 UTC m=+0.069811202
+2023/04/12 13:28:29 Debug T0a: 2023-04-12 13:28:29.240750182 +0000 UTC m=+12.442633158
+2023/04/12 13:29:25 Debug T1: 2023-04-12 13:29:25.616568767 +0000 UTC m=+76.761274236
+2023/04/12 13:29:27 Debug T2: 2023-04-12 13:29:27.025945415 +0000 UTC m=+78.170650884
+2023/04/12 13:30:26 
+2023/04/12 13:30:26 ============================================================================
+2023/04/12 13:30:26 Debug T0: 2023-04-12 13:30:26.142976219 +0000 UTC m=+0.058277231
+2023/04/12 13:30:38 Debug T0a: 2023-04-12 13:30:38.503939039 +0000 UTC m=+12.419240151
+2023/04/12 13:31:49 
+2023/04/12 13:31:49 ============================================================================
+2023/04/12 13:31:49 Debug T0: 2023-04-12 13:31:49.106635939 +0000 UTC m=+0.025093294
+2023/04/12 13:31:51 Debug T1: 2023-04-12 13:31:51.63967304 +0000 UTC m=+222.784378609
+2023/04/12 13:32:02 Debug T0a: 2023-04-12 13:32:02.275835522 +0000 UTC m=+13.194292777
+2023/04/12 13:32:05 Debug T2: 2023-04-12 13:32:05.972717527 +0000 UTC m=+237.117422996
+2023/04/12 13:32:16 Debug T3: 2023-04-12 13:32:16.798259862 +0000 UTC m=+120.374659705
+2023/04/12 13:32:40 Debug T1: 2023-04-12 13:32:40.674068746 +0000 UTC m=+271.818774215
+2023/04/12 13:32:54 Debug T2: 2023-04-12 13:32:54.026559213 +0000 UTC m=+285.171264782
+2023/04/12 13:33:06 Debug T3: 2023-04-12 13:33:06.895265186 +0000 UTC m=+170.471665129
+2023/04/13 01:25:30 
+2023/04/13 01:25:30 
+2023/04/13 01:25:30 T0 2023-04-13 01:25:30.867676002 +0000 UTC m=+0.026911101
+2023/04/13 01:25:49 T0a 2023-04-13 01:25:49.278697865 +0000 UTC m=+18.437932964
+2023/04/13 01:26:37 
+2023/04/13 01:26:37 
+2023/04/13 01:26:37 T1 2023-04-13 01:26:37.36591502 +0000 UTC m=+98.536959377
+2023/04/13 01:26:42 T2 2023-04-13 01:26:42.689456682 +0000 UTC m=+103.860501139
+2023/04/13 01:26:49 T3 2023-04-13 01:26:49.503966744 +0000 UTC m=+100.245536026
