+++ example/lvldbdata/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-000061
+MANIFEST-000087
