+++ example/lvldbdata/CURRENT.bak
@@ -1,1 +1,1 @@
-MANIFEST-000058
+MANIFEST-000084
